backend_django/requirements.txt
django>=5.0
djangorestframework
django-cors-headers
python-socketio[asgi]
uvicorn
gunicorn
python-dotenv
asgiref
djangorestframework-simplejwt
google-generativeai
